Remove size field from messagecontainer
Add more validation on oni_threadEscape, oni_threadRestore
Add cppcheck, will need to remove unusedFunction,missingInclude to hush up warnings
Add crash call stack
Add MSGPACK to all fileexplorer messages

Fix cleanup of RPC clients when they disconnect
Fix null pointer dereference in messagemanager_sendRequest
Fix UAF in fileexplorer_read_callback when sending error response
Fix FileExplorerStat message (missing LSpare)
Fix TimeSpec structure

Update oni-framework README.md

+++ Tools/MiraUtils/Client/TimeSpec.cs
@@ -5,7 +5,7 @@ namespace MiraUtils.Client
     public class TimeSpec : MessageSerializable
     {
         public long Seconds;
-        public int NanoSeconds;
+        public long NanoSeconds;
 
         public TimeSpec()
         {
@@ -21,7 +21,7 @@ namespace MiraUtils.Client
         public override void Deserialize(BinaryReader p_Reader)
         {
             Seconds = p_Reader.ReadInt64();
-            NanoSeconds = p_Reader.ReadInt32();
+            NanoSeconds = p_Reader.ReadInt64();
         }
 
         public override byte[] Serialize()
